Add different mui icon, add new input to be an array
